% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupliHunter.R
\name{dupliHunter}
\alias{dupliHunter}
\title{dupliHunter: Find Similar Duplicates}
\usage{
dupliHunter(data, title_col, author_col, doi_col, threshold = 0.03)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{title_col}{The name of the column containing titles.}

\item{author_col}{The name of the column containing author names.}

\item{doi_col}{The name of the column containing DOIs.}

\item{threshold}{The similarity threshold for titles (default is 0.03).}
}
\value{
A data frame with an additional column 'Duplicate' indicating duplicates.
}
\description{
This function identifies similar duplicates in a dataset based on titles and author names.
}
